+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0](https://github.com/aleksbelic/romanize-deromanize/releases/tag/2.1.0) - 2026-04-09
+
+### Added
+
+- new integration tests for both require and import usage
+
+### Changed
+
+- updates npm dev dependencies
+- migrates ESlint config
+- migrates Jest tests
+- enhances deromanize function validation
+
+### Fixed
+
+- added missing `&&` operator in package.json script
+
 ## [2.0.2](https://github.com/aleksbelic/romanize-deromanize/releases/tag/2.0.2) - 2023-11-05
 
 ### Changed